add source moved next to source drop down instead of inside
