+++ components/Controls/Source.vue
@@ -18,18 +18,9 @@ div.flex.grow.relative
 			ListboxOptions(
 				class="max-h-[600px]"
 			).md_max-h-96.absolute.top-14.z-30.bg-white.rounded-xl.w-full.max-h-96.divide-y.divide-cupboard-lg.overflow-y-auto.overscroll-contain.drop-shadow-standard
+				refreshNewSource
 				ListboxOption(v-for="source in sources" @click="handleSource(source.name)").p-1.text-center.text-xl.cursor-pointer.text-wrap.hover_bg-cupboard-lg
 					| {{ source.name }}
-				.div.div.flex.items-center.gap-2.p-2.border-t.mt-2
-					input(
-						placeholder="Enter new source"
-						type="text"
-						v-model="newSource"
-						@keydown.enter="addSource"
-						@keydown.space="allowSpace"
-					).flex-1.p-2.border.rounded-lg.outline-none
-					button(@click="addSource").bg-utd-green.text-white.rounded-full.w-12.h-12.flex.place-content-center.place-items-center.hover_drop-shadow-standard
-						| +
 </template>
 
 <script lang="ts" setup>
@@ -39,7 +30,6 @@ import { Listbox, ListboxButton, ListboxOption, ListboxOptions } from "@headless
 const emit = defineEmits(["sourceChange"])
 
 const selectedSource = ref("")
-const newSource = ref("")
 
 const { data: sources, refresh } = await useFetch("/api/controls/sources")
 
@@ -51,18 +41,8 @@ const handleSource = (sourceName) => {
 	}
 }
 
-const addSource = async () => {
-	await $fetch("/api/controls/sources", {
-		method: "PUT",
-		body: JSON.stringify({ source: newSource.value }),
-	})
-	newSource.value = ""
+const refreshNewSource = () => {
 	refresh()
-}
-
-// spaces cant be entered normally for some reason so heres a quick fix
-const allowSpace = (event: KeyboardEvent) => {
-	newSource.value += " "
 }
 
 // emit source changes
