+++ components/InventoryReviewChanges.vue
@@ -1,6 +1,15 @@
 <template lang="pug">
 div.flex.flex-col.pt-2.pb-5.px-5.overflow-y-auto.overscroll-contain
-	ControlsSource(@sourceChange="(selectedSource) => (source = selectedSource)").max-md_order-first.max-md_pb-3
+	container.flex
+		ControlsSource(@sourceChange="(selectedSource) => (source = selectedSource)").max-md_order-first.max-md_pb-3
+		input(
+			placeholder="Enter new source"
+			type="text"
+			v-model="newSource"
+			@keydown.enter="addSource"
+		).flex-1.p-2.border.rounded-lg.outline-none
+		button(@click="addSource").bg-utd-green.text-white.rounded-full.w-12.h-10.flex.place-content-center.place-items-center.hover_drop-shadow-standard
+			PlusIcon.fill-white.stroke-white.h-6
 	div.divide-y.divide-cupboard-lg.mb-5
 		div(v-for="(change, index) in props.changes").py-2
 			InventoryReviewItemCard(:change="change" :id="index")
@@ -13,12 +22,24 @@ div.flex.flex-col.pt-2.pb-5.px-5.overflow-y-auto.overscroll-contain
 </template>
 
 <script lang="ts" setup>
+import { PlusIcon } from "@heroicons/vue/24/solid"
+
+const newSource = ref("")
+
 const props = defineProps({
 	changes: {
 		type: Object,
 		required: true,
 	},
 })
+
+const addSource = async () => {
+	await $fetch("/api/controls/sources", {
+		method: "PUT",
+		body: JSON.stringify({ source: newSource.value }),
+	})
+	newSource.value = ""
+}
 
 const emit = defineEmits(["cancel", "accept", "sourceChange"])
 const source = ref("")
